RAG/rag_bot/requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
python-dotenv
faiss-cpu
openai-whisper
pytube
